fix: Update cancel_most_recent_operation to use OperationManager
- Fix cancel_most_recent_operation to delegate to OperationManager instead of accessing active_operations directly
- Add missing cancel_most_recent_operation method to OperationManager
- Ensures proper separation of concerns in the refactored architecture
