+++ crates/semantic-search-client/src/client/async_implementation.rs
@@ -451,29 +451,7 @@ impl AsyncSemanticSearchClient {
 
     /// Cancel the most recent operation
     pub async fn cancel_most_recent_operation(&self) -> Result<String> {
-        let operations = self.active_operations.read().await;
-
-        if operations.is_empty() {
-            return Err(SemanticSearchError::OperationFailed(
-                "No active operations to cancel".to_string(),
-            ));
-        }
-
-        // Find the most recent operation (highest started_at time)
-        let most_recent = operations
-            .iter()
-            .max_by_key(|(_, handle)| handle.started_at)
-            .map(|(id, _)| *id);
-
-        drop(operations); // Release the read lock
-
-        if let Some(operation_id) = most_recent {
-            self.cancel_operation(operation_id).await
-        } else {
-            Err(SemanticSearchError::OperationFailed(
-                "No active operations found".to_string(),
-            ))
-        }
+        self.operation_manager.cancel_most_recent_operation().await
     }
 
     /// Cancel all active operations
